reorganize files and update to camel style
