+++ stadium/Assets/Scripts/ColorControl.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2e84672b857237f47876489d54251048
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
