--- stadium/Assets/Scripts/LEDEffects.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bd58da5ecaa39eb4b88bd4b6fb196ab3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
