--- stadium/Assets/Scripts/buttonController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0f1d1f2afb05d494f971fd72418ce83d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
