--- stadium/Assets/Scripts/cameraControls.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3aa83a1bbaa14b64d9aef70946871c32
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
