--- stadium/Assets/Scripts/colorButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 849415f0b8d7d78408e7a15127edce6f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
