--- stadium/Assets/Scripts/inputReader.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a93a3d54d92082044b296672e0905b5d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
